Commit 4
Commit 4

+++ src/br/com/ygor/controller/CursoController.java
@@ -35,7 +35,10 @@ public class CursoController implements Controller {
     
     @Override
     public void delete(){
-        dao.delete(view.getId());
+        if (!dao.hasRelationshipWithAnotherTable(view.getId()))
+            dao.delete(view.getId());
+        else
+            view.msgExclusionNotAllowed();
     }
     
     @Override
